chore: update next.js
