+++ packages/takudev-web/package.json
@@ -32,13 +32,13 @@
     "defaults"
   ],
   "dependencies": {
-    "next": "12.1.6",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
+    "next": "12.2.3",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-fast-compare": "^3.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.2",
+    "@babel/core": "^7.18.9",
     "@storybook/addon-actions": "^6.5.9",
     "@storybook/addon-essentials": "^6.5.9",
     "@storybook/addon-interactions": "^6.5.9",
@@ -50,32 +50,44 @@
     "@storybook/testing-library": "^0.0.13",
     "@stylelint/postcss-css-in-js": "^0.38.0",
     "@types/babel__core": "^7.1.18",
-    "@types/react": "18.0.8",
-    "@types/react-dom": "18.0.3",
+    "@types/eslint": "^8.4.5",
+    "@types/jest": "^28.1.6",
+    "@types/node": "18.0.6",
+    "@types/react": "18.0.15",
+    "@types/react-dom": "18.0.6",
     "@types/throttle-debounce": "^5.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.30.7",
+    "@typescript-eslint/parser": "^5.30.7",
     "babel-loader": "^8.2.3",
     "babel-plugin-react-require": "^3.1.3",
+    "buffer": "^6.0.3",
+    "date-fns": "^2.29.1",
     "esbuild": "^0.14.21",
     "esbuild-register": "^3.3.2",
-    "eslint-config-next": "12.1.6",
+    "eslint": "^8.20.0",
+    "eslint-config-next": "12.2.3",
+    "eslint-import-resolver-typescript": "^3.3.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-storybook": "^0.6.1",
     "postcss": "8",
     "postcss-jsx": "^0.36.4",
     "postcss-scss": "^4.0.4",
     "postcss-syntax": "^0.36.2",
-    "puppeteer": "^15.4.0",
+    "puppeteer": "^15.5.0",
     "reg-keygen-git-hash-plugin": "^0.12.1",
     "reg-notify-github-plugin": "^0.12.1",
     "reg-publish-s3-plugin": "^0.12.1",
     "reg-suit": "^0.12.1",
-    "sass": "^1.52.1",
+    "sass": "^1.54.0",
     "sass-loader": "^10.0.0",
     "storycap": "^3.1.8",
     "stylelint": "^14.8.5",
     "stylelint-config-recess-order": "^3.0.0",
     "stylelint-config-standard-scss": "^4.0.0",
     "stylelint-scss": "^4.2.0",
-    "throttle-debounce": "^5.0.0"
+    "throttle-debounce": "^5.0.0",
+    "ts-jest": "^28.0.7",
+    "turbo": "^1.3.4"
   },
   "packageManager": "yarn@1.22.18"
 }
