+++ packages/takudev-web/pages/api/basic.ts
@@ -1,0 +1,7 @@
+import type { NextApiRequest, NextApiResponse } from 'next';
+
+export default function handler(_: NextApiRequest, res: NextApiResponse) {
+  res.statusCode = 401;
+  res.setHeader('WWW-authenticate', 'Basic realm="Secure Area"');
+  res.end('Basic Auth Required');
+}
